Removed nested index.js files and moved them to their parent directory with the name of the folder they were nested in. Moved the client test file into the routers directory. Removed index.js file nested within the routes directory. Changed file url references to files that were moved.
