--- api/routers/index.js
@@ -1,5 +1,0 @@
-exports.auth = require('./auth');
-exports.roles = require('./roles');
-exports.clients = require('./clients');
-exports.jobsheets = require('./jobsheets');
-exports.projects = require('./projects');
